ADD global internal error + pointer error functions
error codes will now be tracked in a global variable
called g_error_internal. This makes it possible to check it at any time,
and to set it without having to chain return values, or use pointers.
Also added 2 version of the handle_error functions that will take and
return a pointer. This is useful in functions that need to return but
don't return an int. it can also be cast to and from a void * if it is
something different from a void pointer.
also improved text wrapping in function comments in the header.