add builtins tests and improve variable types/perm
completed the builtins, added tests for all builtins,
and improved how we handle variable types and permissions
in the env list. Each variable now has a bitstock of options,
which define what type of variable it is and if it is read only.
modified relevant functions to work with this.