fix multiline commands breaking history
when a a history entry contains a newline,
this used to break history since the history would split on newlines,
unexpectedly splitting that command into 2 commands.
we now split on the history prefix (:INDEX:), which avoids this issue.
to avoid issues with commands containing the prefix, the split also
checks if there is a newline or the beginning of the history file before
the history prefix.