move the env list into the shell struct and more
the env list now lives in the shell struct so it is easier to access
from different parts of the shell. (hence the many changed files.)
History now also handles HISTFILE and HOME as expected,
and i made a start for the init_shell function.
this still needs to initialize wanted shell variables, though.