improve the content of oldpwd when cwd is removed
for example, if the cwd is `/tmp/somedir`, and `/tmp/somedir` is
removed, the content of OLDPWD would be emptied if we cd'd away from
that directory. now the cd builtin will look at the env var PWD
if, AND ONLY IF, it can not manually retrieve the current directory.