fix exit not working as last command in a pipeline
the return code of execute_pipe was incorrect since we were returning
the exit code of the command before the pipe symbol, instead of the exit
code of the command after the pipe symbol. fixing this also fixed exit
not working if it is the last comand in a pipeline since the return code
of the exit builtin needs to be passed back to the cetushell function.