fix "exit" being printed at the wrong time
when exiting using ctrl+d, "exit" was not printed, and when exit was the
last command in a pipeline, and the first command took more time, "exit"
would already be printed before that first command was done.
now "exit" is printed at the very end, making sure it is always printed
when the shell exits.