fix bug in heredoc and improve heredoc handling
quotes were removed in heredocs, while they should not be. expansions
were also only done if they were not in quotes, but in heredocs
expansions should always be done as quote characters are ignored.
ctrl+d now also works in heredocs to terminate the input (as if the EOF
was entered).