fix errors in expansion in heredocs
if an environment variable contains a backslash, but the variable was
contained in single quotes in a heredoc, an extra backslash was printed.
now any weird combination of quotes and expansion should be handled.
get_env and expand_variable have gotten some extra options to handle
this correctly. also added some tests for these cases.