rebuild word processing
the process_word function and all quoting functions that it used are now
completely rebuilt, making it a single function to complete all tasks.
it is implemented as finite state machine that can take a table_type to
specify which transition table should be used, so the correct patterns
are used for regular words, heredocs, or here end words.
this also means that extra tables can be added easily if needed in the
future, and existing behaviour can be changed without chaning the code,
since only the transition table needs to be updated.