implement backslash completion and improve prompt
an unterminated backslash (input is unquoted but ends with a backslash),
is now completed with complete_quote, and a relevant prompt is shown.
The normal prompt now uses the shell variable PS1, and initially it is
set to a coloured version of the prompt if TERM=xterm-256color.
also simplified / cleaned up some code for quoting.