fix bug were closing an fd did not work
when trying to close an fd with the syntax 0>&-, the given fd (here 0)
would be overwritten again if there was another redirection after it,
essentially undoing the closure.
To solve this, we now keep track of fd's we want to close, but only
actually close them once we've handled all other redirections and no
other redirections for this fd have been given.
this means we still respect the order of all redirections, but also
ensure closing an fd works as expected.