WIP of job_control
terminal ownership implemented, and basic concepts for job control
in the executor seem to be working, but no testing done yet.
We also don't have the command strings for the jobs/processes yet,
and the job list leaks for now, as it is never used nor freed.
we also do not have any way to track jobs yet.