fix bug with unterminated backslash
if an unterminated backslash was complete, we did not insert a newline.
This would cause any special character
following it (for example a quote) to lose its meaning.
We now ALWAYS insert a newline when completing any form of quotes,
so the backslash escapes the newline and it is removed,
without affecting following characters.