fix bug where previous command ID was incorrect
the logic that updated the previous and current command was incomplete,
causing it to be wrong in certain cases. To fix this i repurposed
the get_new_job_id function to get a correct previous id,
and added a static function in check_jobs that has complete logic
for determining the current and previous job id after a job is removed.