add jobs builtin and job_spec_match
added a first version of the jobs builtin, which should be feature
but is not yet using the getopts function, and is not yet norm compliant
also added job_spec_match, a standalone function to check a given job
to a given job_spec string.
this can be used in all job control builtins.
the structure and output of other functions that print job output have
been updated aswell, so everyting is more modular and logical.