add bg builtin and tests
added the bg builtin and tests.
bg does not take options, and simply processes every argument
as a job_spec, sending every process in every job matching
that job_spec a SIGCONT to continue the process.
If the job was previously flagged as foreground, bg also sets the
foreground flag to false and adds a '&' to the command string.
status printing is left to check_jobs to simplify this builtin.