fix job_control not working in non-interactive
even though fg will not work in non-interactive mode for obvious reasons
jobs should work, and there is no reason why bg could not function.
this is updated now, but unfortunately I had to add a debug flag to the
job_control struct. we can remove this when we turn the project in.