add pipe completion
if a pipe is incomplete, we now ask for extra input to complete it.
this is done in the same manner as with quotes, but I needed to add some
ugly exceptions in the tokenizer, mostly for when you hit return right
after the pipe symbol, since with quotes the added input would become
part of the same token als the quote symbol, while with pipes it should
always become a new token. also added a test for pipe completion.