remove setting status based on g_error_code
We don't really use g_error_code, and setting the status code from it if
it is non-zero can create really weird results, since the user only
expect the most recent EXIT code to be in the status code ($?).
So I decided to not process it at all. It can still be used for
debugging, but is not used other than that.