fix not detecting commands exiting because of sig
for example, if a program segfaults, we did not print anything.
this can be tested with:
python -c 'import os, signal;os.kill(os.getpid(), signal.SIGSEGV)'
we now check for signals right after executing a command,
and when checking job statuses. If the command was part of a pipe,
we also print the command as text, to make clear which command it was.