fix invalid read on ctrl + right when cursor index is 0
To fix this, the movement behavior was replaced with that used on TOSH.
The main difference is that TOSH places the cursor on the left or right
side of the word, depending on which direction the cursor was moved,
while Cetushell always placed the cursor on the left side. The TOSH
behavior was also the same as that of bash and fish.